Refactor rendering pipeline: reduce duplication and improve module boundaries
- Remove dead code in record_decode.rs: delete unused extract_sweep_data(),
  inline decode_record_to_radials_timed into its only caller, remove DecodeTimings
- Extract CPU azimuth search into find_nearest_azimuth_index() helper, deduplicating
  value_at_polar() and collection_time_at_polar() in gpu_renderer.rs
- Pass &Renderers to render_canvas_with_geo instead of 5 individual Option params
- Extract color table/LUT code into new color_table.rs module (~275 lines of pure
  functions moved out of gpu_renderer.rs)
- Extract compute_gpu_sweep_state() helper from canvas.rs inline logic
- Parameterize GLSL shader functions: replace 4 duplicated azimuth lookup functions
  (find_nearest_az/find_nearest_prev_az/find_bracket_az/find_bracket_prev_az) and
  2 duplicated sample functions with 3 parameterized versions that accept sampler2D

Net reduction: ~415 lines. No behavior changes.
