+++ src/nexrad/gpu_renderer.rs
@@ -3,276 +3,13 @@
 //! Renders polar radar data (azimuths x gates) directly on the GPU using a fragment
 //! shader that performs polar-to-Cartesian conversion and color lookup from a LUT texture.
 
+use super::color_table::{
+    build_reflectivity_lut, continuous_color_scale, product_from_str, product_value_range,
+};
 use crate::state::RenderProcessing;
 use glow::HasContext;
-use nexrad_render::{Color as NrColor, ColorScale, ColorStop, ContinuousColorScale, Product};
+use nexrad_render::Product;
 use std::sync::Arc;
-
-// Default value ranges per product (used for color LUT normalization).
-pub fn product_value_range(product: Product) -> (f32, f32) {
-    match product {
-        Product::Reflectivity => (-32.0, 95.0),
-        Product::Velocity => (-64.0, 64.0),
-        Product::SpectrumWidth => (0.0, 30.0),
-        Product::DifferentialReflectivity => (-2.0, 6.0),
-        Product::CorrelationCoefficient => (0.0, 1.05),
-        Product::DifferentialPhase => (0.0, 360.0),
-        Product::ClutterFilterPower => (-20.0, 20.0),
-    }
-}
-
-fn product_from_str(s: &str) -> Product {
-    match s {
-        "velocity" => Product::Velocity,
-        "spectrum_width" => Product::SpectrumWidth,
-        "differential_reflectivity" => Product::DifferentialReflectivity,
-        "differential_phase" => Product::DifferentialPhase,
-        "correlation_coefficient" => Product::CorrelationCoefficient,
-        "clutter_filter_power" => Product::ClutterFilterPower,
-        _ => Product::Reflectivity,
-    }
-}
-
-// --- OKLab color space helpers for perceptually uniform interpolation ---
-
-fn srgb_to_linear(c: f32) -> f32 {
-    if c <= 0.04045 {
-        c / 12.92
-    } else {
-        ((c + 0.055) / 1.055).powf(2.4)
-    }
-}
-
-fn linear_to_srgb(c: f32) -> f32 {
-    if c <= 0.0031308 {
-        c * 12.92
-    } else {
-        1.055 * c.powf(1.0 / 2.4) - 0.055
-    }
-}
-
-/// Convert sRGB (0-1) to OKLab (L, a, b).
-#[allow(clippy::excessive_precision)]
-fn srgb_to_oklab(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
-    let lr = srgb_to_linear(r);
-    let lg = srgb_to_linear(g);
-    let lb = srgb_to_linear(b);
-
-    let l = 0.4122214708 * lr + 0.5363325363 * lg + 0.0514459929 * lb;
-    let m = 0.2119034982 * lr + 0.6806995451 * lg + 0.1073969566 * lb;
-    let s = 0.0883024619 * lr + 0.2817188376 * lg + 0.6299787005 * lb;
-
-    let l_ = l.cbrt();
-    let m_ = m.cbrt();
-    let s_ = s.cbrt();
-
-    (
-        0.2104542553 * l_ + 0.7936177850 * m_ - 0.0040720468 * s_,
-        1.9779984951 * l_ - 2.4285922050 * m_ + 0.4505937099 * s_,
-        0.0259040371 * l_ + 0.7827717662 * m_ - 0.8086757660 * s_,
-    )
-}
-
-/// Convert OKLab (L, a, b) to sRGB (0-1), clamped.
-#[allow(clippy::excessive_precision)]
-fn oklab_to_srgb(ol: f32, oa: f32, ob: f32) -> (f32, f32, f32) {
-    let l_ = ol + 0.3963377774 * oa + 0.2158037573 * ob;
-    let m_ = ol - 0.1055613458 * oa - 0.0638541728 * ob;
-    let s_ = ol - 0.0894841775 * oa - 1.2914855480 * ob;
-
-    let l = l_ * l_ * l_;
-    let m = m_ * m_ * m_;
-    let s = s_ * s_ * s_;
-
-    let r = 4.0767416621 * l - 3.3077115913 * m + 0.2309699292 * s;
-    let g = -1.2684380046 * l + 2.6097574011 * m - 0.3413193965 * s;
-    let b = -0.0041960863 * l - 0.7034186147 * m + 1.7076147010 * s;
-
-    (
-        linear_to_srgb(r.clamp(0.0, 1.0)),
-        linear_to_srgb(g.clamp(0.0, 1.0)),
-        linear_to_srgb(b.clamp(0.0, 1.0)),
-    )
-}
-
-fn smoothstep(edge0: f32, edge1: f32, x: f32) -> f32 {
-    let t = ((x - edge0) / (edge1 - edge0)).clamp(0.0, 1.0);
-    t * t * (3.0 - 2.0 * t)
-}
-
-/// Build a 1024-entry RGBA LUT for reflectivity using OKLab interpolation,
-/// monotonically increasing luminance, and a low-end alpha ramp.
-pub fn build_reflectivity_lut(min_val: f32, max_val: f32) -> Vec<u8> {
-    // Anchor colors: (dBZ, r, g, b, a) in sRGB 0-1.
-    // Designed for black background with increasing luminance.
-    let anchors: &[(f32, f32, f32, f32, f32)] = &[
-        (0.0, 0.00, 0.00, 0.00, 0.00),  // black
-        (5.0, 0.10, 0.10, 0.14, 0.15),  // near-black, faint
-        (10.0, 0.20, 0.22, 0.32, 0.40), // dim blue-grey
-        (15.0, 0.35, 0.40, 0.58, 0.75), // slate blue
-        (20.0, 0.15, 0.72, 0.15, 0.90), // SHARP bright green
-        (28.0, 0.05, 0.35, 0.08, 1.00), // dark green
-        (32.0, 0.90, 0.88, 0.10, 1.00), // SHARP bright yellow
-        (37.0, 0.68, 0.64, 0.10, 1.00), // duller yellow
-        (40.0, 0.92, 0.58, 0.08, 1.00), // SHARP bright orange
-        (45.0, 0.70, 0.45, 0.06, 1.00), // warm dark yellow-orange
-        (50.0, 0.85, 0.12, 0.10, 1.00), // SHARP bright red
-        (55.0, 0.52, 0.10, 0.08, 1.00), // dark red
-        (60.0, 0.92, 0.68, 0.72, 1.00), // SHARP blush
-        (65.0, 0.95, 0.25, 0.55, 1.00), // hot pink
-        (70.0, 0.68, 0.20, 0.85, 1.00), // SHARP bright purple
-        (75.0, 0.28, 0.08, 0.40, 1.00), // dark purple
-        (80.0, 0.20, 0.82, 0.85, 1.00), // SHARP bright cyan
-        (90.0, 0.08, 0.30, 0.34, 1.00), // dark cyan
-    ];
-
-    // Pre-convert anchors to OKLab
-    let oklab_anchors: Vec<(f32, f32, f32, f32, f32)> = anchors
-        .iter()
-        .map(|&(dbz, r, g, b, a)| {
-            let (ol, oa, ob) = srgb_to_oklab(r, g, b);
-            (dbz, ol, oa, ob, a)
-        })
-        .collect();
-
-    let lut_size = 1024usize;
-    let mut lut_data = Vec::with_capacity(lut_size * 4);
-    let range = max_val - min_val;
-
-    for i in 0..lut_size {
-        let t = i as f32 / (lut_size - 1) as f32;
-        // Optional perceptual shaping: slightly expand midrange
-        let t_shaped = t.powf(0.9);
-        let dbz = min_val + t_shaped * range;
-
-        // Find bracketing anchors
-        let mut seg = 0usize;
-        #[allow(clippy::needless_range_loop)]
-        for j in 1..oklab_anchors.len() {
-            if oklab_anchors[j].0 >= dbz {
-                seg = j - 1;
-                break;
-            }
-            seg = j - 1;
-        }
-        let lo = &oklab_anchors[seg];
-        let hi = if seg + 1 < oklab_anchors.len() {
-            &oklab_anchors[seg + 1]
-        } else {
-            lo
-        };
-
-        let frac = if (hi.0 - lo.0).abs() > 0.001 {
-            ((dbz - lo.0) / (hi.0 - lo.0)).clamp(0.0, 1.0)
-        } else {
-            0.0
-        };
-
-        // Interpolate in OKLab space
-        let ol = lo.1 + frac * (hi.1 - lo.1);
-        let oa = lo.2 + frac * (hi.2 - lo.2);
-        let ob = lo.3 + frac * (hi.3 - lo.3);
-        let alpha_anchor = lo.4 + frac * (hi.4 - lo.4);
-
-        let (r, g, b) = oklab_to_srgb(ol, oa, ob);
-
-        // Low-end alpha ramp: smoothstep from 3 to 12 dBZ
-        let alpha_ramp = smoothstep(3.0, 12.0, dbz);
-        let alpha = (alpha_anchor * alpha_ramp).clamp(0.0, 1.0);
-
-        lut_data.push((r * 255.0).round() as u8);
-        lut_data.push((g * 255.0).round() as u8);
-        lut_data.push((b * 255.0).round() as u8);
-        lut_data.push((alpha * 255.0).round() as u8);
-    }
-
-    lut_data
-}
-
-/// Build a continuous (linearly-interpolated) color scale for the given product.
-/// Uses the same color stops as the discrete NWS scales from the nexrad_render crate,
-/// but interpolates smoothly between them for a commercial-quality appearance.
-pub fn continuous_color_scale(product: Product) -> ColorScale {
-    let stops = match product {
-        Product::Reflectivity => vec![
-            // Unused for LUT generation (build_reflectivity_lut is used instead),
-            // but kept as fallback for color_scale.color() calls.
-            ColorStop::new(0.0, NrColor::rgba(0.0000, 0.0000, 0.0000, 0.0)),
-            ColorStop::new(5.0, NrColor::rgba(0.04, 0.14, 0.18, 0.15)),
-            ColorStop::new(15.0, NrColor::rgba(0.08, 0.35, 0.30, 0.70)),
-            ColorStop::new(25.0, NrColor::rgba(0.22, 0.56, 0.16, 0.95)),
-            ColorStop::new(35.0, NrColor::rgb(0.62, 0.64, 0.10)),
-            ColorStop::new(45.0, NrColor::rgb(0.82, 0.48, 0.10)),
-            ColorStop::new(55.0, NrColor::rgb(0.80, 0.14, 0.16)),
-            ColorStop::new(65.0, NrColor::rgb(0.70, 0.22, 0.56)),
-            ColorStop::new(75.0, NrColor::rgb(0.88, 0.65, 0.80)),
-        ],
-        Product::Velocity => vec![
-            ColorStop::new(-64.0, NrColor::rgb(0.00, 0.15, 0.00)), // very dark green
-            ColorStop::new(-50.0, NrColor::rgb(0.00, 0.38, 0.00)), // dark green
-            ColorStop::new(-36.0, NrColor::rgb(0.00, 0.65, 0.00)), // medium green
-            ColorStop::new(-26.0, NrColor::rgb(0.00, 0.90, 0.00)), // bright green
-            ColorStop::new(-16.0, NrColor::rgb(0.45, 0.95, 0.35)), // light green
-            ColorStop::new(-5.0, NrColor::rgb(0.68, 0.78, 0.68)),  // gray-green
-            ColorStop::new(0.0, NrColor::rgb(0.60, 0.60, 0.60)),   // neutral gray
-            ColorStop::new(5.0, NrColor::rgb(0.78, 0.68, 0.68)),   // gray-red
-            ColorStop::new(16.0, NrColor::rgb(0.95, 0.40, 0.35)),  // light red
-            ColorStop::new(26.0, NrColor::rgb(0.90, 0.00, 0.00)),  // bright red
-            ColorStop::new(36.0, NrColor::rgb(0.65, 0.00, 0.00)),  // medium red
-            ColorStop::new(50.0, NrColor::rgb(0.40, 0.00, 0.00)),  // dark red
-            ColorStop::new(64.0, NrColor::rgb(0.18, 0.00, 0.00)),  // very dark red
-        ],
-        Product::SpectrumWidth => vec![
-            ColorStop::new(0.0, NrColor::rgb(0.5020, 0.5020, 0.5020)),
-            ColorStop::new(4.0, NrColor::rgb(0.0000, 0.0000, 0.8039)),
-            ColorStop::new(8.0, NrColor::rgb(0.0000, 0.8039, 0.8039)),
-            ColorStop::new(12.0, NrColor::rgb(0.0000, 0.8039, 0.0000)),
-            ColorStop::new(16.0, NrColor::rgb(0.9333, 0.9333, 0.0000)),
-            ColorStop::new(20.0, NrColor::rgb(1.0000, 0.6471, 0.0000)),
-            ColorStop::new(25.0, NrColor::rgb(1.0000, 0.0000, 0.0000)),
-        ],
-        Product::DifferentialReflectivity => vec![
-            ColorStop::new(-2.0, NrColor::rgb(0.5020, 0.0000, 0.5020)),
-            ColorStop::new(-1.0, NrColor::rgb(0.0000, 0.0000, 0.8039)),
-            ColorStop::new(0.0, NrColor::rgb(0.6627, 0.6627, 0.6627)),
-            ColorStop::new(0.5, NrColor::rgb(0.5647, 0.9333, 0.5647)),
-            ColorStop::new(1.5, NrColor::rgb(0.9333, 0.9333, 0.0000)),
-            ColorStop::new(2.5, NrColor::rgb(1.0000, 0.6471, 0.0000)),
-            ColorStop::new(4.0, NrColor::rgb(1.0000, 0.0000, 0.0000)),
-        ],
-        Product::CorrelationCoefficient => vec![
-            ColorStop::new(0.0, NrColor::rgb(0.0000, 0.0000, 0.0000)),
-            ColorStop::new(0.2, NrColor::rgb(0.3922, 0.0000, 0.5882)),
-            ColorStop::new(0.5, NrColor::rgb(0.0000, 0.0000, 0.8039)),
-            ColorStop::new(0.7, NrColor::rgb(0.0000, 0.5451, 0.5451)),
-            ColorStop::new(0.85, NrColor::rgb(0.0000, 0.8039, 0.4000)),
-            ColorStop::new(0.92, NrColor::rgb(0.0000, 0.8039, 0.0000)),
-            ColorStop::new(0.96, NrColor::rgb(0.9333, 0.9333, 0.0000)),
-            ColorStop::new(0.98, NrColor::rgb(0.9020, 0.9020, 0.9020)),
-        ],
-        Product::DifferentialPhase => vec![
-            ColorStop::new(0.0, NrColor::rgb(0.5020, 0.0000, 0.5020)),
-            ColorStop::new(45.0, NrColor::rgb(0.0000, 0.0000, 0.8039)),
-            ColorStop::new(90.0, NrColor::rgb(0.0000, 0.8039, 0.8039)),
-            ColorStop::new(135.0, NrColor::rgb(0.0000, 0.8039, 0.0000)),
-            ColorStop::new(180.0, NrColor::rgb(0.9333, 0.9333, 0.0000)),
-            ColorStop::new(225.0, NrColor::rgb(1.0000, 0.6471, 0.0000)),
-            ColorStop::new(270.0, NrColor::rgb(1.0000, 0.0000, 0.0000)),
-            ColorStop::new(315.0, NrColor::rgb(1.0000, 0.0000, 1.0000)),
-        ],
-        Product::ClutterFilterPower => vec![
-            ColorStop::new(-20.0, NrColor::rgb(0.0000, 0.0000, 0.5451)),
-            ColorStop::new(-10.0, NrColor::rgb(0.0000, 0.0000, 0.8039)),
-            ColorStop::new(-5.0, NrColor::rgb(0.6784, 0.8471, 0.9020)),
-            ColorStop::new(-1.0, NrColor::rgb(0.6627, 0.6627, 0.6627)),
-            ColorStop::new(1.0, NrColor::rgb(0.6627, 0.6627, 0.6627)),
-            ColorStop::new(5.0, NrColor::rgb(1.0000, 0.7529, 0.7961)),
-            ColorStop::new(10.0, NrColor::rgb(1.0000, 0.4118, 0.4118)),
-            ColorStop::new(20.0, NrColor::rgb(0.8039, 0.0000, 0.0000)),
-        ],
-    };
-    ColorScale::Continuous(ContinuousColorScale::new(stops))
-}
 
 const VERTEX_SHADER: &str = r#"#version 300 es
 precision highp float;
@@ -340,26 +77,16 @@ uniform float u_prev_max_range_km;
 
 const float PI = 3.14159265359;
 
-// Sample the raw data texture at a given (gate_index, azimuth_index).
-// Returns 0.0 for out-of-range (sentinel: below threshold).
-float sample_data(float g, float a) {
-    if (g < 0.0 || g >= u_gate_count || a < 0.0 || a >= u_azimuth_count) {
+// Sample a data texture at a given (gate_index, azimuth_index).
+// Parameterized by sampler and dimensions so the same function works for
+// both current and previous sweep textures.
+float sample_data_p(sampler2D data_tex, float gate_count, float azimuth_count, float g, float a) {
+    if (g < 0.0 || g >= gate_count || a < 0.0 || a >= azimuth_count) {
         return 0.0;
     }
-    float gu = (g + 0.5) / u_gate_count;
-    float av = (a + 0.5) / u_azimuth_count;
-    return texture(u_data_tex, vec2(gu, av)).r;
-}
-
-// Sample the previous scan's data texture at a given (gate_index, azimuth_index).
-// Uses prev-texture dimensions for correct UV mapping when sweeps differ.
-float sample_prev_data(float g, float a) {
-    if (g < 0.0 || g >= u_prev_gate_count || a < 0.0 || a >= u_prev_azimuth_count) {
-        return 0.0;
-    }
-    float gu = (g + 0.5) / u_prev_gate_count;
-    float av = (a + 0.5) / u_prev_azimuth_count;
-    return texture(u_prev_data_tex, vec2(gu, av)).r;
+    float gu = (g + 0.5) / gate_count;
+    float av = (a + 0.5) / azimuth_count;
+    return texture(data_tex, vec2(gu, av)).r;
 }
 
 // Raw values 0 (below threshold) and 1 (range folded) are sentinels.
@@ -368,20 +95,20 @@ bool is_valid(float v) {
 }
 
 // Find the nearest azimuth index for a given angle in degrees.
+// Parameterized by azimuth count and sampler.
 // Returns -1.0 if no radial is close enough (gap).
-// Also writes the azimuth angle of the found radial to out_az.
-float find_nearest_az(float azimuth_deg, out float out_az) {
-    float az_spacing = 360.0 / u_azimuth_count;
+float find_nearest_az_p(float azimuth_deg, float az_count, sampler2D az_tex, out float out_az) {
+    float az_spacing = 360.0 / az_count;
     float est_idx = azimuth_deg / az_spacing;
-    float inv_count = 1.0 / u_azimuth_count;
+    float inv_count = 1.0 / az_count;
 
     float best_idx = 0.0;
     float best_dist = 360.0;
     float best_az = 0.0;
 
     for (float offset = -2.0; offset <= 2.0; offset += 1.0) {
-        float i = floor(mod(est_idx + offset, u_azimuth_count));
-        float tex_az = texture(u_azimuth_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
+        float i = floor(mod(est_idx + offset, az_count));
+        float tex_az = texture(az_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
         float d = abs(azimuth_deg - tex_az);
         d = min(d, 360.0 - d);
         if (d < best_dist) {
@@ -391,7 +118,6 @@ float find_nearest_az(float azimuth_deg, out float out_az) {
         }
     }
 
-    // Gap detection
     if (best_dist > az_spacing * 1.5) {
         out_az = 0.0;
         return -1.0;
@@ -401,98 +127,20 @@ float find_nearest_az(float azimuth_deg, out float out_az) {
 }
 
 // Find the two nearest azimuth indices that bracket the given angle.
+// Parameterized by azimuth count and sampler.
 // Returns false if in a gap region.
-bool find_bracket_az(float azimuth_deg, out float idx_lo, out float idx_hi, out float frac) {
-    float az_spacing = 360.0 / u_azimuth_count;
+bool find_bracket_az_p(float azimuth_deg, float az_count, sampler2D az_tex,
+                       out float idx_lo, out float idx_hi, out float frac) {
+    float az_spacing = 360.0 / az_count;
     float est_idx = azimuth_deg / az_spacing;
-    float inv_count = 1.0 / u_azimuth_count;
-
-    // Collect candidates (up to 5)
-    float cand_idx[5];
-    float cand_az[5];
-    for (int k = 0; k < 5; k++) {
-        float i = floor(mod(est_idx + float(k - 2), u_azimuth_count));
-        cand_idx[k] = i;
-        cand_az[k] = texture(u_azimuth_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
-    }
-
-    // Find the closest radial on each side (or equal)
-    float lo_idx = -1.0, hi_idx = -1.0;
-    float lo_az = -999.0, hi_az = 999.0;
-    float lo_dist = 360.0, hi_dist = 360.0;
-
-    for (int k = 0; k < 5; k++) {
-        float az = cand_az[k];
-        // Signed angular difference (target - candidate), wrapped to -180..180
-        float diff = azimuth_deg - az;
-        diff = mod(diff + 540.0, 360.0) - 180.0;
-
-        if (diff >= 0.0 && diff < lo_dist) {
-            lo_dist = diff;
-            lo_idx = cand_idx[k];
-            lo_az = az;
-        }
-        if (diff <= 0.0 && (-diff) < hi_dist) {
-            hi_dist = -diff;
-            hi_idx = cand_idx[k];
-            hi_az = az;
-        }
-    }
-
-    if (lo_idx < 0.0 || hi_idx < 0.0) return false;
-
-    // Gap check
-    float span = lo_dist + hi_dist;
-    if (span > az_spacing * 1.5) return false;
-
-    idx_lo = lo_idx;
-    idx_hi = hi_idx;
-    frac = (span > 0.001) ? lo_dist / span : 0.0;
-    return true;
-}
-
-// Previous-sweep azimuth lookup (mirrors find_nearest_az using prev textures).
-float find_nearest_prev_az(float azimuth_deg, out float out_az) {
-    float az_spacing = 360.0 / u_prev_azimuth_count;
-    float est_idx = azimuth_deg / az_spacing;
-    float inv_count = 1.0 / u_prev_azimuth_count;
-
-    float best_idx = 0.0;
-    float best_dist = 360.0;
-    float best_az = 0.0;
-
-    for (float offset = -2.0; offset <= 2.0; offset += 1.0) {
-        float i = floor(mod(est_idx + offset, u_prev_azimuth_count));
-        float tex_az = texture(u_prev_azimuth_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
-        float d = abs(azimuth_deg - tex_az);
-        d = min(d, 360.0 - d);
-        if (d < best_dist) {
-            best_dist = d;
-            best_idx = i;
-            best_az = tex_az;
-        }
-    }
-
-    if (best_dist > az_spacing * 1.5) {
-        out_az = 0.0;
-        return -1.0;
-    }
-    out_az = best_az;
-    return best_idx;
-}
-
-// Previous-sweep bracket lookup (mirrors find_bracket_az using prev textures).
-bool find_bracket_prev_az(float azimuth_deg, out float idx_lo, out float idx_hi, out float frac) {
-    float az_spacing = 360.0 / u_prev_azimuth_count;
-    float est_idx = azimuth_deg / az_spacing;
-    float inv_count = 1.0 / u_prev_azimuth_count;
+    float inv_count = 1.0 / az_count;
 
     float cand_idx[5];
     float cand_az[5];
     for (int k = 0; k < 5; k++) {
-        float i = floor(mod(est_idx + float(k - 2), u_prev_azimuth_count));
+        float i = floor(mod(est_idx + float(k - 2), az_count));
         cand_idx[k] = i;
-        cand_az[k] = texture(u_prev_azimuth_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
+        cand_az[k] = texture(az_tex, vec2((i + 0.5) * inv_count, 0.5)).r;
     }
 
     float lo_idx = -1.0, hi_idx = -1.0;
@@ -549,7 +197,7 @@ void main() {
         desat_factor = clamp((age - 0.75) / 0.25, 0.0, 1.0) * 0.9;
     }
 
-    // --- Unified sweep pipeline: select spatial params based on use_prev ---
+    // --- Unified sweep pipeline: select spatial params and samplers based on use_prev ---
     float s_gate_count    = use_prev ? u_prev_gate_count    : u_gate_count;
     float s_azimuth_count = use_prev ? u_prev_azimuth_count : u_azimuth_count;
     float s_first_gate_km = use_prev ? u_prev_first_gate_km : u_first_gate_km;
@@ -570,15 +218,21 @@ void main() {
         return;
     }
 
-    // Azimuth lookup + value sampling
+    // Select the active samplers for the chosen sweep (current or previous).
+    // GLSL ES 3.00 requires sampler access in uniform control flow, but
+    // use_prev is uniform across the draw call for a given pixel's branch —
+    // we use explicit if/else to satisfy the compiler.
     float value;
 
     if (u_interpolation == 1) {
         // Bilinear interpolation
         float az_lo, az_hi, az_frac;
-        bool bracket_ok = use_prev
-            ? find_bracket_prev_az(azimuth_deg, az_lo, az_hi, az_frac)
-            : find_bracket_az(azimuth_deg, az_lo, az_hi, az_frac);
+        bool bracket_ok;
+        if (use_prev) {
+            bracket_ok = find_bracket_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, az_lo, az_hi, az_frac);
+        } else {
+            bracket_ok = find_bracket_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, az_lo, az_hi, az_frac);
+        }
         if (!bracket_ok) {
             fragColor = vec4(0.0);
             return;
@@ -590,15 +244,15 @@ void main() {
 
         float v00, v10, v01, v11;
         if (use_prev) {
-            v00 = sample_prev_data(g_lo, az_lo);
-            v10 = sample_prev_data(g_hi, az_lo);
-            v01 = sample_prev_data(g_lo, az_hi);
-            v11 = sample_prev_data(g_hi, az_hi);
+            v00 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_lo, az_lo);
+            v10 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_hi, az_lo);
+            v01 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_lo, az_hi);
+            v11 = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, g_hi, az_hi);
         } else {
-            v00 = sample_data(g_lo, az_lo);
-            v10 = sample_data(g_hi, az_lo);
-            v01 = sample_data(g_lo, az_hi);
-            v11 = sample_data(g_hi, az_hi);
+            v00 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_lo, az_lo);
+            v10 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_hi, az_lo);
+            v01 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_lo, az_hi);
+            v11 = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, g_hi, az_hi);
         }
 
         // Weighted average skipping sentinel values
@@ -622,16 +276,21 @@ void main() {
     } else {
         // Nearest neighbor
         float dummy_az;
-        float best_idx = use_prev
-            ? find_nearest_prev_az(azimuth_deg, dummy_az)
-            : find_nearest_az(azimuth_deg, dummy_az);
+        float best_idx;
+        if (use_prev) {
+            best_idx = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, dummy_az);
+        } else {
+            best_idx = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, dummy_az);
+        }
         if (best_idx < 0.0) {
             fragColor = vec4(0.0);
             return;
         }
-        value = use_prev
-            ? sample_prev_data(floor(gate_idx), best_idx)
-            : sample_data(floor(gate_idx), best_idx);
+        if (use_prev) {
+            value = sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, floor(gate_idx), best_idx);
+        } else {
+            value = sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, floor(gate_idx), best_idx);
+        }
     }
 
     if (!is_valid(value)) {
@@ -642,9 +301,12 @@ void main() {
     // Despeckle filter
     if (u_despeckle_enabled == 1) {
         float dummy_az2;
-        float center_az = use_prev
-            ? find_nearest_prev_az(azimuth_deg, dummy_az2)
-            : find_nearest_az(azimuth_deg, dummy_az2);
+        float center_az;
+        if (use_prev) {
+            center_az = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_prev_azimuth_tex, dummy_az2);
+        } else {
+            center_az = find_nearest_az_p(azimuth_deg, s_azimuth_count, u_azimuth_tex, dummy_az2);
+        }
         float center_g = floor(gate_idx);
         if (center_az >= 0.0) {
             int valid_count = 0;
@@ -654,9 +316,9 @@ void main() {
                     float ng = center_g + float(dg);
                     float na = mod(center_az + float(da), s_azimuth_count);
                     if (use_prev) {
-                        if (is_valid(sample_prev_data(ng, na))) valid_count++;
+                        if (is_valid(sample_data_p(u_prev_data_tex, s_gate_count, s_azimuth_count, ng, na))) valid_count++;
                     } else {
-                        if (is_valid(sample_data(ng, na))) valid_count++;
+                        if (is_valid(sample_data_p(u_data_tex, s_gate_count, s_azimuth_count, ng, na))) valid_count++;
                     }
                 }
             }
@@ -690,6 +352,36 @@ void main() {
     fragColor = vec4(color.rgb * a, a);
 }
 "#;
+
+/// Find the nearest azimuth index in an array of azimuth angles.
+///
+/// Returns `None` if the nearest azimuth is farther than 1.5x the expected spacing
+/// (gap detection). Used by CPU-side inspector lookups.
+fn find_nearest_azimuth_index(
+    azimuths: &[f32],
+    azimuth_count: usize,
+    target_deg: f32,
+) -> Option<usize> {
+    let mut best_idx = 0usize;
+    let mut best_dist = 360.0f32;
+    for (i, &az) in azimuths.iter().enumerate() {
+        let mut d = (target_deg - az).abs();
+        if d > 180.0 {
+            d = 360.0 - d;
+        }
+        if d < best_dist {
+            best_dist = d;
+            best_idx = i;
+        }
+    }
+
+    let az_spacing = 360.0 / azimuth_count as f32;
+    if best_dist > az_spacing * 1.5 {
+        return None;
+    }
+
+    Some(best_idx)
+}
 
 /// GPU-based radar renderer using WebGL2 shaders.
 pub struct RadarGpuRenderer {
@@ -1266,52 +958,32 @@ impl RadarGpuRenderer {
             return None;
         }
 
-        // Check range bounds
         if range_km < self.first_gate_km || range_km >= self.max_range_km {
             return None;
         }
 
-        // Find nearest azimuth index
-        let az_count = self.azimuth_count as usize;
+        let az_idx = find_nearest_azimuth_index(
+            &self.cpu_azimuths,
+            self.azimuth_count as usize,
+            azimuth_deg,
+        )?;
+
         let gate_count = self.gate_count as usize;
-        let mut best_idx = 0usize;
-        let mut best_dist = 360.0f32;
-        for (i, &az) in self.cpu_azimuths.iter().enumerate() {
-            let mut d = (azimuth_deg - az).abs();
-            if d > 180.0 {
-                d = 360.0 - d;
-            }
-            if d < best_dist {
-                best_dist = d;
-                best_idx = i;
-            }
-        }
-
-        // Gap check: if nearest azimuth is too far away, no data
-        let az_spacing = 360.0 / az_count as f32;
-        if best_dist > az_spacing * 1.5 {
-            return None;
-        }
-
-        // Compute gate index
         let gate_idx = ((range_km - self.first_gate_km) / self.gate_interval_km).floor() as usize;
         if gate_idx >= gate_count {
             return None;
         }
 
-        // Data layout: row-major [azimuth][gate]
-        let offset = best_idx * gate_count + gate_idx;
+        let offset = az_idx * gate_count + gate_idx;
         if offset >= self.cpu_gate_values.len() {
             return None;
         }
 
         let raw = self.cpu_gate_values[offset];
-        // Raw sentinels: 0 = below threshold, 1 = range folded
         if raw <= 1.0 {
             return None;
         }
 
-        // Convert raw to physical value
         if self.data_scale == 0.0 {
             Some(raw)
         } else {
@@ -1327,26 +999,13 @@ impl RadarGpuRenderer {
             return None;
         }
 
-        let az_count = self.azimuth_count as usize;
-        let mut best_idx = 0usize;
-        let mut best_dist = 360.0f32;
-        for (i, &az) in self.cpu_azimuths.iter().enumerate() {
-            let mut d = (azimuth_deg - az).abs();
-            if d > 180.0 {
-                d = 360.0 - d;
-            }
-            if d < best_dist {
-                best_dist = d;
-                best_idx = i;
-            }
-        }
+        let az_idx = find_nearest_azimuth_index(
+            &self.cpu_azimuths,
+            self.azimuth_count as usize,
+            azimuth_deg,
+        )?;
 
-        let az_spacing = 360.0 / az_count as f32;
-        if best_dist > az_spacing * 1.5 {
-            return None;
-        }
-
-        self.cpu_radial_times.get(best_idx).copied()
+        self.cpu_radial_times.get(az_idx).copied()
     }
 
     /// Sweep-aware value lookup for the inspector.
